// src/types/JobSearchInfo.ts
type JobSearchInfo = {
  jobList: string;
  id: string;
  Status: string;
  name: string;
  owner: string;
  created: string;
  historyStatus: string;
  applicationName: number;
  lineItemCount: string;
  errorItemCount: string;
  lastUpdated: string;
  lastRun: string;
  runId: string;
  starred: string;
  schedule: string;
  historyPercentage: string;
  nextScheduleDt: string;
};

export default JobSearchInfo;